% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bank_1min}
\alias{bank_1min}
\title{Bank index 1 Minute Chart data}
\format{A data frame with 8 variables: \code{index}, \code{date},
\code{time}, \code{open}, \code{high}, \code{low} , \code{close} and \code{id}}
\usage{
bank_1min
}
\description{
Bank Nifty represents the 12 most liquid and
large capitalised stocks from the banking sector
which trade on the National Stock Exchange (NSE). 
It provides investors and market intermediaries a 
benchmark that captures the capital market performance of Indian banking sector.
}
\keyword{datasets}

