% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{companyList}
\alias{companyList}
\title{companyList for stocks}
\format{A data frame with 3 variables: \code{company}, \code{id},
and  \code{file}}
\usage{
companyList
}
\description{
It contains the 502 company name list and their id for accessing the 
datasource. Company performance interms of 1 Min chart data can be further 
loaded from their corresponding file name.
}
\keyword{datasets}

