% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nifty_1min}
\alias{nifty_1min}
\title{NIFTY50 1 Minute Chart data}
\format{A data frame with 8 variables: \code{index}, \code{date},
\code{time}, \code{open}, \code{high}, \code{low} , \code{close} and \code{id}}
\usage{
nifty_1min
}
\description{
The NIFTY 50 index is National Stock Exchange of India's benchmark 
stock market index for Indian equity market.
It is a well diversified 50 stock index accounting for 22 sectors 
of the economy. It is used for a variety of purposes such as benchmarking 
fund portfolios, index based derivatives and index funds.
}
\keyword{datasets}

