% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nse_BB}
\alias{nse_BB}
\title{Individual stocks 1 Minute Chart data}
\format{A data frame with five variables: \code{id}, \code{date},
\code{time}, \code{open}, \code{high}, \code{low} , \code{close} and \code{vol}}
\usage{
nse_BB
}
\description{
Choosing individual stocks from other companies can have 
advantages over fund allocation for the investors. 
Predicting an individual stock price is not a easy task,
while it depends on many external sources such as company performance,
government policy, public expectation and media focus, etc.
Stock price are considered to be
very dynamic and some dependency with technical indicators.
These datasets are aggregated with their alphabetical order (groups)
}
\keyword{datasets}

